backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas>=2.2.0
numpy>=1.26.0
google-generativeai>=0.5.4
python-dotenv==1.0.1
pydantic==2.7.1
openpyxl==3.1.2
httpx==0.27.0
